requirements.txt
streamlit==1.32.0
spotipy==2.25.1
feedparser==6.0.11
requests==2.32.3
tqdm==4.67.1
python-dotenv==1.0.1
huggingface_hub==0.24.1 